refactor: isolate clawhub skill search
